GCW-3133 addressed PR comments
