+++ app/controllers/add_user.js
@@ -1,8 +1,10 @@
 import config from "stock/config/environment";
+import AsyncMixin, { ERROR_STRATEGIES } from "stock/mixins/async";
+import ImageUploadMixin from "stock/mixins/image_upload";
 import Ember from "ember";
 const { getOwner } = Ember;
 
-export default Ember.Controller.extend({
+export default Ember.Controller.extend(AsyncMixin, ImageUploadMixin, {
   queryParams: ["organisationId"],
   organisationId: null,
   organisationName: null,
@@ -93,8 +95,6 @@ export default Ember.Controller.extend({
     let image = this.get("newUploadedImage");
     if (image) {
       return image.save();
-    } else {
-      return Promise.resolve();
     }
   },
 
@@ -102,7 +102,7 @@ export default Ember.Controller.extend({
     searchOrganization() {
       this.replaceRoute("search_organisation", {
         queryParams: {
-          backToNewUser: true
+          redirectToPath: "add_user"
         }
       });
     },
@@ -120,42 +120,21 @@ export default Ember.Controller.extend({
     },
 
     saveUser() {
-      let _this = this;
-      const loadingView = getOwner(_this)
-        .lookup("component:loading")
-        .append();
+      let newUser = this.get("store").createRecord(
+        "user",
+        this.getRequestParams()
+      );
 
-      let newUser = _this
-        .get("store")
-        .createRecord("user", _this.getRequestParams());
+      newUser.set("userRoleIds", this.getWithDefault("selectedRoleIds", []));
 
-      if (_this.get("selectedRoleIds.length")) {
-        newUser.set("userRoleIds", _this.get("selectedRoleIds"));
-      } else {
-        newUser.set("userRoleIds", []);
-      }
+      return this.runTask(async () => {
+        newUser.set("image", await this.saveImage());
 
-      _this.saveImage().then(imageData => {
-        if (imageData) {
-          let image = _this.get("store").peekRecord("image", imageData.id);
-          newUser.set("image", image);
-        }
+        const { id } = await newUser.save();
 
-        newUser
-          .save()
-          .then(data => {
-            this.clearFormData();
-            this.transitionToRoute("user_details", data.id);
-          })
-          .catch(xhr => {
-            if (xhr.status === 422) {
-              this.get("messageBox").alert(xhr.responseJSON.errors[0].message);
-            } else {
-              throw xhr;
-            }
-          })
-          .finally(() => loadingView.destroy());
-      });
+        this.clearFormData();
+        this.transitionToRoute("user_details", id);
+      }, ERROR_STRATEGIES.MODAL);
     },
 
     cancelForm() {
@@ -174,108 +153,6 @@ export default Ember.Controller.extend({
         },
         this.locale("no").string
       );
-    },
-
-    uploadReady() {
-      this.set("isReady", true);
-    },
-
-    uploadStart(e, data) {
-      this.set("uploadedFileDate", data);
-      Ember.$(".loading-image-indicator").show();
-      this.set("loadingPercentage", "Image Uploading ");
-    },
-
-    cancelUpload() {
-      if (this.get("uploadedFileDate")) {
-        this.get("uploadedFileDate").abort();
-      }
-    },
-
-    uploadProgress(e, data) {
-      e.target.disabled = true; // disable image-selection
-      var progress = parseInt((data.loaded / data.total) * 100, 10) || 0;
-      this.set("addPhotoLabel", progress + "%");
-      this.set("loadingPercentage", "Image Uploading " + progress + "%");
-    },
-
-    uploadComplete(e) {
-      e.target.disabled = false; // enable image-selection
-      this.set("uploadedFileDate", null);
-      Ember.$(".loading-image-indicator.hide_image_loading").hide();
-      this.set("loadingPercentage", "Image Uploading ");
-    },
-
-    uploadSuccess(e, data) {
-      var identifier =
-        data.result.version +
-        "/" +
-        data.result.public_id +
-        "." +
-        data.result.format;
-      var newUploadedImage = this.get("store").createRecord("image", {
-        cloudinaryId: identifier,
-        favourite: true
-      });
-      this.set("newUploadedImage", newUploadedImage);
-      this.set("userImageKeys", identifier);
-    },
-
-    initActionSheet: function(onSuccess) {
-      return window.plugins.actionsheet.show(
-        {
-          buttonLabels: [
-            this.locale("edit_images.upload").string,
-            this.locale("edit_images.camera").string,
-            this.locale("edit_images.cancel").string
-          ]
-        },
-        function(buttonIndex) {
-          if (buttonIndex === 1) {
-            navigator.camera.getPicture(onSuccess, null, {
-              quality: 40,
-              destinationType: navigator.camera.DestinationType.DATA_URL,
-              sourceType: navigator.camera.PictureSourceType.PHOTOLIBRARY
-            });
-          }
-          if (buttonIndex === 2) {
-            navigator.camera.getPicture(onSuccess, null, {
-              correctOrientation: true,
-              quality: 40,
-              destinationType: navigator.camera.DestinationType.DATA_URL,
-              sourceType: navigator.camera.PictureSourceType.CAMERA
-            });
-          }
-          if (buttonIndex === 3) {
-            window.plugins.actionsheet.hide();
-          }
-        }
-      );
-    },
-
-    triggerUpload() {
-      // For Cordova application
-      if (config.cordova.enabled) {
-        var onSuccess = (function() {
-          return function(path) {
-            console.log(path);
-            var dataURL = "data:image/jpg;base64," + path;
-
-            Ember.$("input[type='file']").fileupload(
-              "option",
-              "formData"
-            ).file = dataURL;
-            Ember.$("input[type='file']").fileupload("add", {
-              files: [dataURL]
-            });
-          };
-        })(this);
-
-        this.initActionSheet(onSuccess);
-      } else {
-        // For web application
-        Ember.$("input[type='file']").trigger("click");
-      }
     }
   }
 });
