+++ app/routes/search_organisation.js
@@ -2,6 +2,6 @@ import AuthorizeRoute from "./authorize";
 
 export default AuthorizeRoute.extend({
   queryParams: {
-    backToNewUser: false
+    redirectToPath: null
   }
 });
