+++ app/templates/user_details.hbs
@@ -142,4 +142,3 @@
 
   </div>
 </section>
-{{partial "loading_image"}}
